requirements.txt
discord.py==1.7.3
python-decouple==3.4
PyNaCl==1.4.0
dnspython==2.1.0
async-timeout==3.0.1
Cogs==0.4.4